refactor(editor): studio panel contracts 분리
